pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "storycast-seed-agents"
version = "0.1.0"
description = "Autonomous multimodal storytelling agent for the Seed Agents Challenge"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "Rohan5commit" }
]
dependencies = [
  "byteplus-python-sdk-v2>=0.3.17",
  "fastapi>=0.115.0",
  "httpx>=0.28.0",
  "jinja2>=3.1.4",
  "pydantic>=2.9.2",
  "python-dotenv>=1.0.1",
  "tenacity>=9.0.0",
  "uvicorn[standard]>=0.32.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3"
]

[project.scripts]
storycast = "storycast.cli:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["storycast*"]

[tool.setuptools.package-data]
storycast = [
  "web/templates/*.html",
  "web/static/*.css",
  "web/static/*.js"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
